rename new-classes to classes and fix imports
