+++ lib/classes/employee.js
