+++ lib/classes/intern.js
