+++ lib/classes/manager.js
