+++ lib/inquirer-prompts/createEngineer.js
@@ -1,4 +1,4 @@
-const Engineer = require('../new-classes/engineer')
+const Engineer = require('../classes/engineer')
 
 async function createEngineer() {
 	let newEngineer = await inquirer.prompt([
