+++ lib/inquirer-prompts/createIntern.js
@@ -1,4 +1,4 @@
-const Intern = require('../new-classes/intern')
+const Intern = require('../classes/intern')
 
 async function createIntern() {
 	let newIntern = await inquirer.prompt([
