+++ lib/inquirer-prompts/createManager.js
@@ -1,4 +1,4 @@
-const Manager = require('../new-classes/manager')
+const Manager = require('../classes/manager')
 
 async function createManager() {
 	let newManager = await inquirer.prompt([
